SaveState: Simplify and convert to std::thread
