+++ pcsx2/pcsx2.vcxproj
@@ -697,8 +697,6 @@
     </ClCompile>
     <ClCompile Include="gui\Saveslots.cpp" />
     <ClCompile Include="gui\SysState.cpp" />
-    <ClCompile Include="ZipTools\thread_gzip.cpp" />
-    <ClCompile Include="ZipTools\thread_lzma.cpp" />
     <ClCompile Include="windows\Optimus.cpp" />
   </ItemGroup>
   <ItemGroup>
@@ -1089,7 +1087,6 @@
     <ClInclude Include="gui\RecentIsoList.h" />
     <ClInclude Include="PathDefs.h" />
     <ClInclude Include="SysForwardDefs.h" />
-    <ClInclude Include="ZipTools\ThreadedZipTools.h" />
     <ClInclude Include="cheatscpp.h" />
   </ItemGroup>
   <ItemGroup>
